requirements.txt
tesseract
easyocr


streamlit
pymupdf
pdfplumber
sentence-transformers
faiss-cpu
matplotlib
plotly
requests
pandas
numpy
pytesseract
pillow
torch